+++ src/Netsphere.Server.Chat/Mailbox.cs
@@ -106,7 +106,7 @@ namespace Netsphere.Server.Chat
                 {
                     Id = _idGeneratorService.GetNextId(IdKind.Mail),
                     PlayerId = account.Id,
-                    SenderPlayerId = (int)Player.Account.Id,
+                    SenderPlayerId = (long)Player.Account.Id,
                     SentDate = DateTimeOffset.Now.ToUnixTimeSeconds(),
                     Title = title,
                     Message = message,
